Handle unit test cases failure Signed-off-by: maheshlokhande1 <mahesh.lokhande@ibm.com>
